requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# AI and LLM integration
openai>=1.0.0
anthropic>=0.7.0
crewai>=0.1.0
langchain>=0.0.350
langchain-openai>=0.0.5

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.1
pdfplumber>=0.9.0

# JIRA integration
jira>=3.5.0
atlassian-python-api>=3.41.0

# HTTP requests and API clients
requests>=2.31.0
httpx>=0.25.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
pathlib2>=2.3.7
typing-extensions>=4.8.0
python-dateutil>=2.8.2

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# MCP (Model Context Protocol) - if available
# mcp>=0.1.0  # Uncomment when MCP package is available

# Optional: Enhanced document processing
# textract>=1.6.5  # For advanced document extraction
# spacy>=3.7.0     # For NLP processing
